Fix: use parseFloar instead of parseint in Number validator

+++ src/types/number.js
@@ -2,7 +2,7 @@ module.exports = {
 	allowedProps: ['required', 'enum', 'coerce'],
 
 	validateType(value, schema = {}) {
-		if (schema.coerce === true) value = parseInt(value);
+		if (schema.coerce === true) value = parseFloat(value);
 		return typeof value === 'number' && !Number.isNaN(value);
 	}
 };
